Change forgery protection to :null_session
Every time rails apps create a form, an authenticity token is randomly generated. It then gets passed to the controller when the form is submitted. This is to prevent someone from making an outside request that didn't start at our website. Normally we throw an exception by setting this to `:exception`. But in the case of API, we WANT to let requests through since there is no form that's loaded and no authenticity token generated. We're more or less saying "We'll handle this ourselves"